# src/build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"ApiResponse_any_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {},
					"message": {
						"type": "string"
					},
					"error": {
						"type": "string"
					},
					"details": {},
					"pagination": {
						"properties": {
							"totalPages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"totalPages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"DailySummaryResponse": {
				"properties": {
					"summary_id": {
						"type": "string"
					},
					"user_id": {
						"type": "string"
					},
					"summary_date": {
						"type": "string",
						"format": "date-time"
					},
					"content": {
						"type": "string"
					}
				},
				"required": [
					"summary_id",
					"user_id",
					"summary_date",
					"content"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_DailySummaryResponse-Array_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/DailySummaryResponse"
						},
						"type": "array"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"type": "string"
					},
					"details": {},
					"pagination": {
						"properties": {
							"totalPages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"totalPages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DailySummaryListRequest": {
				"properties": {
					"user_id": {
						"type": "string"
					},
					"summary_date": {
						"type": "string",
						"format": "date-time"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_DailySummaryResponse_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"$ref": "#/components/schemas/DailySummaryResponse"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"type": "string"
					},
					"details": {},
					"pagination": {
						"properties": {
							"totalPages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"totalPages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateDailySummaryRequest": {
				"properties": {
					"user_id": {
						"type": "string"
					},
					"summary_date": {
						"type": "string",
						"format": "date-time"
					},
					"content": {
						"type": "string"
					}
				},
				"required": [
					"user_id",
					"summary_date",
					"content"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateDailySummaryRequest": {
				"properties": {
					"summary_id": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"summary_date": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"summary_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CourseScheduleResponse": {
				"properties": {
					"plan_id": {
						"type": "string"
					},
					"user_id": {
						"type": "string"
					},
					"course_id": {
						"type": "string"
					},
					"start_date": {
						"type": "string",
						"format": "date-time"
					},
					"end_date": {
						"type": "string",
						"format": "date-time"
					},
					"status": {
						"type": "string"
					}
				},
				"required": [
					"plan_id",
					"user_id",
					"course_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_CourseScheduleResponse-Array_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/CourseScheduleResponse"
						},
						"type": "array"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"type": "string"
					},
					"details": {},
					"pagination": {
						"properties": {
							"totalPages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"totalPages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_CourseScheduleResponse_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"$ref": "#/components/schemas/CourseScheduleResponse"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"type": "string"
					},
					"details": {},
					"pagination": {
						"properties": {
							"totalPages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"totalPages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateCourseScheduleRequest": {
				"properties": {
					"user_id": {
						"type": "string"
					},
					"course_id": {
						"type": "string"
					},
					"start_date": {
						"type": "string",
						"format": "date-time"
					},
					"end_date": {
						"type": "string",
						"format": "date-time"
					},
					"status": {
						"type": "string"
					}
				},
				"required": [
					"user_id",
					"course_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CourseResponse": {
				"properties": {
					"course_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"icon_url": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"default_ai_persona_id": {
						"type": "string"
					}
				},
				"required": [
					"course_id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_CourseResponse-Array_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/CourseResponse"
						},
						"type": "array"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"type": "string"
					},
					"details": {},
					"pagination": {
						"properties": {
							"totalPages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"totalPages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_CourseResponse_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"$ref": "#/components/schemas/CourseResponse"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"type": "string"
					},
					"details": {},
					"pagination": {
						"properties": {
							"totalPages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"totalPages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateCourseRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"icon_url": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"default_ai_persona_id": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateCourseRequest": {
				"properties": {
					"course_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"icon_url": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"default_ai_persona_id": {
						"type": "string"
					}
				},
				"required": [
					"course_id"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "express-pg-orm-backend-ts",
		"version": "1.0.0",
		"description": "TypeScript Express.js backend with PostgreSQL and Sequelize ORM",
		"contact": {}
	},
	"paths": {
...
		"/dailySummary/list": {
			"post": {
				"operationId": "ListDailySummary",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_DailySummaryResponse-Array_"
								}
							}
						}
					}
				},
				"description": "查询每日总结列表（可选按用户/日期过滤）",
				"tags": [
					"每日总结"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DailySummaryListRequest"
							}
						}
					}
				}
			}
		},
		"/dailySummary/add": {
			"post": {
				"operationId": "AddDailySummary",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_DailySummaryResponse_"
								}
							}
						}
					}
				},
				"description": "新增每日总结",
				"tags": [
					"每日总结"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateDailySummaryRequest"
							}
						}
					}
				}
			}
		},
		"/dailySummary/update": {
			"post": {
				"operationId": "UpdateDailySummary",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"description": "修改每日总结",
				"tags": [
					"每日总结"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateDailySummaryRequest"
							}
						}
					}
				}
			}
		},
		"/dailySummary/delete": {
			"post": {
				"operationId": "DeleteDailySummary",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"description": "删除每日总结",
				"tags": [
					"每日总结"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "string"
							}
						}
					}
				}
			}
		},
		"/course-schedules/list": {
			"post": {
				"operationId": "ListCourseSchedules",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_CourseScheduleResponse-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"课程安排表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"limit": {
										"type": "number",
										"format": "double"
									},
									"page": {
										"type": "number",
										"format": "double"
									}
								},
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/course-schedules/getById": {
			"post": {
				"operationId": "GetCourseScheduleById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_CourseScheduleResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"课程安排表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"plan_id": {
										"type": "string"
									}
								},
								"required": [
									"plan_id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/course-schedules/add": {
			"post": {
				"operationId": "AddCourseSchedule",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"课程安排表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCourseScheduleRequest"
							}
						}
					}
				}
			}
		},
		"/course-schedules/update": {
			"post": {
				"operationId": "UpdateCourseSchedule",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"课程安排表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"plan_id": {
										"type": "string"
									}
								},
								"required": [
									"plan_id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/course-schedules/delete": {
			"post": {
				"operationId": "DeleteCourseSchedule",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"课程安排表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"plan_id": {
										"type": "string"
									}
								},
								"required": [
									"plan_id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/courses/list": {
			"post": {
				"operationId": "ListCourses",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_CourseResponse-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"课程表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"limit": {
										"type": "number",
										"format": "double"
									},
									"page": {
										"type": "number",
										"format": "double"
									}
								},
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/courses/getCourseById": {
			"post": {
				"operationId": "GetCourseById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_CourseResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"课程表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"course_id": {
										"type": "string"
									}
								},
								"required": [
									"course_id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/courses/add": {
			"post": {
				"operationId": "AddCourse",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"课程表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCourseRequest"
							}
						}
					}
				}
			}
		},
		"/courses/update": {
			"post": {
				"operationId": "UpdateCourse",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"课程表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateCourseRequest"
							}
						}
					}
				}
			}
		},
		"/courses/delete": {
			"post": {
				"operationId": "DeleteCourse",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"课程表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"course_id": {
										"type": "string"
									}
								},
								"required": [
									"course_id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}